+++ json_serializable/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 6.12.0
+
+- Support `JsonKey` annotation on constructor parameters.
+- Add support for `JsonSerializable(createJsonSchema: true)`.
+
 ## 6.11.4
 
 - Allow `analyzer: >=9.0.0 <11.0.0`